hv: pci: refine codes to make Service VM can manage bridge
In previous codes, the bridge is owned by hypervsior if there
is pre-launched VM configured. This can cause bug, e.g. the
net card behind the bridge cannot get IP because the write
to bridge's command register to set BusMaster bit is ignored.

This patch is to refine the codes to make the bridge be managed
by Service VM if the device behind the bridge is not assigned to
pre-launched VM. Furthermore, it adds the parent device into struct
pci_pdev to construct PCI hierarchy.

Tracked-On: #8849
Signed-off-by: Yi Sun <yi.y.sun@intel.com>